+++ src/Integer.php
@@ -6,7 +6,7 @@ use InvalidArgumentException;
  * @author Martin Dilling-Hansen <martindilling@gmail.com>
  * @date   2015-02-10
  */
-class Integer implements ValueObject
+class Integer implements ValueObject, Stringable
 {
     /**
      * @var int
@@ -21,7 +21,7 @@ class Integer implements ValueObject
     public function __construct($value)
     {
         if (!is_integer($value)) {
-            throw new InvalidArgumentException('Provided argument [' . $value . '] is not valid!');
+            throw new InvalidArgumentException('Provided argument [' . $value . '] is not a valid integer!');
         }
 
         $this->value = (int) $value;
@@ -55,5 +55,22 @@ class Integer implements ValueObject
     public function __toString()
     {
         return (string) $this->value;
+    }
+
+    /**
+     * @return string
+     */
+    public function toString()
+    {
+        return $this->__toString();
+    }
+
+    /**
+     * @param string $value
+     * @return static
+     */
+    public static function fromString($value)
+    {
+        return new static((integer) $value);
     }
 }